Small in memory optimization
Changed the way data is stored in memory to reduce overhead. By storing
the data in chunks of native object arrays. This reduced the amount of
pointers per single object in the table making it use 4 times less
memory compared to the previous version. This might even be possible to
be split in half by using a native byte array and use utf-8 instead of
utf-16. As the data is split up in chunks when processing the table
every chunk is assigned to an thread to prevent multiple thread to try
to access the same memory object creating low level queue's. This
resulted in doubling the speed of processing the table.

Personal result changes:
max table size: 30m lines => 190m lines (before getting out of memory
error's)
lines processed per ms: 40k lines => 80k - 90k (initial searches tend to
be around the 60k-70k)

TODO:
allow for custom sized string to be stored in the data table
allow for custom sized chunks
update the from disk functions to work with the new system